+++ src/language_server_list.nim
@@ -3,18 +3,26 @@ import nimsumtree/rope
 import misc/[custom_logger, custom_async, util, response, event]
 import scripting_api except DocumentEditor, TextDocumentEditor, AstDocumentEditor
 import text/language/[language_server_base, lsp_types]
-import document
+import document, config_provider
 
 logCategory "language-server-list"
+
+declareSettings LspMergeSettings, "lsp-merge":
+  ## Timeout for LSP requests in milliseconds
+  declare timeout, int, 10000
 
 type
   MergeStrategy = enum First, All, FirstThenTimeout, AnyThenTimeout
   LanguageServerList* = ref object of LanguageServer
+    config: ConfigStore
+    mergeConfig: LspMergeSettings
     languageServers*: seq[LanguageServer]
     timeout: int
 
-proc newLanguageServerList*(): LanguageServerList =
+proc newLanguageServerList*(config: ConfigStore): LanguageServerList =
   var server = new LanguageServerList
+  server.config = config
+  server.mergeConfig = LspMergeSettings.new(server.config)
   return server
 
 proc updateRefetchWorkspaceSymbolsOnQueryChange*(self: LanguageServerList) =
@@ -27,8 +35,12 @@ proc updateRefetchWorkspaceSymbolsOnQueryChange*(self: LanguageServerList) =
 proc addLanguageServer*(self: LanguageServerList, languageServer: LanguageServer): bool =
   if languageServer in self.languageServers:
     return false
+
+  proc configPriority(languageServer: LanguageServer): int =
+    return self.config.get("lsp." & languageServer.name & ".priority", 0)
+
   self.languageServers.add(languageServer)
-  self.languageServers.sort((a, b) => cmp(a.priority, b.priority))
+  self.languageServers.sort((a, b) => cmp(configPriority(b), configPriority(a)))
   return true
 
 proc removeLanguageServer*(self: LanguageServerList, languageServer: LanguageServer): bool =
@@ -38,15 +50,16 @@ proc removeLanguageServer*(self: LanguageServerList, languageServer: LanguageSer
     return true
   return false
 
-template merge(T: untyped, subCall: untyped, name: untyped): untyped =
+template merge(self: LanguageServerList, T: untyped, subCall: untyped, name: untyped): untyped =
   block:
+    let timeout = self.mergeConfig.timeout.get()
     var futs = newSeq[Future[seq[T]]]()
     var futsTimeout = newSeq[Future[bool]]()
     for lss in self.languageServers:
       let ls {.inject.} = lss
       let fut = subCall
       futs.add fut
-      futsTimeout.add fut.withTimeout(500.milliseconds)
+      futsTimeout.add fut.withTimeout(timeout.milliseconds)
 
     await allFutures(futsTimeout)
 
@@ -57,15 +70,16 @@ template merge(T: untyped, subCall: untyped, name: untyped): untyped =
 
     res
 
-template mergeOption(T: untyped, subCall: untyped, name: untyped): untyped =
+template mergeOption(self: LanguageServerList, T: untyped, subCall: untyped, name: untyped): untyped =
   block:
+    let timeout = self.mergeConfig.timeout.get()
     var futs = newSeq[Future[Option[T]]]()
     var futsTimeout = newSeq[Future[bool]]()
     for lss in self.languageServers:
       let ls {.inject.} = lss
       let fut = subCall
       futs.add fut
-      futsTimeout.add fut.withTimeout(500.milliseconds)
+      futsTimeout.add fut.withTimeout(timeout.milliseconds)
 
     await allFutures(futsTimeout)
 
@@ -78,15 +92,16 @@ template mergeOption(T: untyped, subCall: untyped, name: untyped): untyped =
 
     res
 
-template mergeResponse(T: untyped, subCall: untyped, name: untyped): untyped =
+template mergeResponse(self: LanguageServerList, T: untyped, subCall: untyped, name: untyped): untyped =
   block:
+    let timeout = self.mergeConfig.timeout.get()
     var futs = newSeq[Future[Response[seq[T]]]]()
     var futsTimeout = newSeq[Future[bool]]()
     for lss in self.languageServers:
       let ls {.inject.} = lss
       let fut = subCall
       futs.add fut
-      futsTimeout.add fut.withTimeout(500.milliseconds)
+      futsTimeout.add fut.withTimeout(timeout.milliseconds)
 
     await allFutures(futsTimeout)
 
@@ -108,60 +123,61 @@ method disconnect*(self: LanguageServerList, document: Document) {.gcsafe, raise
     ls.disconnect(document)
 
 method getDefinition*(self: LanguageServerList, filename: string, location: Cursor): Future[seq[Definition]] {.async.} =
-  return merge(Definition, ls.getDefinition(filename, location), "getDefinition")
+  return self.merge(Definition, ls.getDefinition(filename, location), "getDefinition")
 
 method getDeclaration*(self: LanguageServerList, filename: string, location: Cursor): Future[seq[Definition]] {.async.} =
-  return merge(Definition, ls.getDeclaration(filename, location), "getDeclaration")
+  return self.merge(Definition, ls.getDeclaration(filename, location), "getDeclaration")
 
 method getImplementation*(self: LanguageServerList, filename: string, location: Cursor): Future[seq[Definition]] {.async.} =
-  return merge(Definition, ls.getImplementation(filename, location), "getImplementation")
+  return self.merge(Definition, ls.getImplementation(filename, location), "getImplementation")
 
 method getTypeDefinition*(self: LanguageServerList, filename: string, location: Cursor): Future[seq[Definition]] {.async.} =
-  return merge(Definition, ls.getTypeDefinition(filename, location), "getTypeDefinition")
+  return self.merge(Definition, ls.getTypeDefinition(filename, location), "getTypeDefinition")
 
 method getReferences*(self: LanguageServerList, filename: string, location: Cursor): Future[seq[Definition]] {.async.} =
-  return merge(Definition, ls.getReferences(filename, location), "getReferences")
+  return self.merge(Definition, ls.getReferences(filename, location), "getReferences")
 
 method switchSourceHeader*(self: LanguageServerList, filename: string): Future[Option[string]] {.async.} =
-  return mergeOption(string, ls.switchSourceHeader(filename), "switchSourceHeader")
+  return self.mergeOption(string, ls.switchSourceHeader(filename), "switchSourceHeader")
 
 method getCompletions*(self: LanguageServerList, filename: string, location: Cursor): Future[Response[lsp_types.CompletionList]] {.async.} =
   # completions are handled by the completion provider as separate providers, so no need to implement this right now
   discard
 
 method getSymbols*(self: LanguageServerList, filename: string): Future[seq[Symbol]] {.async.} =
-  return merge(Symbol, ls.getSymbols(filename), "getSymbols")
+  return self.merge(Symbol, ls.getSymbols(filename), "getSymbols")
 
 method getWorkspaceSymbols*(self: LanguageServerList, filename: string, query: string): Future[seq[Symbol]] {.async.} =
-  return merge(Symbol, ls.getWorkspaceSymbols(filename, query), "getWorkspaceSymbols")
+  return self.merge(Symbol, ls.getWorkspaceSymbols(filename, query), "getWorkspaceSymbols")
 
 method getHover*(self: LanguageServerList, filename: string, location: Cursor): Future[Option[string]] {.async.} =
-  return mergeOption(string, ls.getHover(filename, location), "getHover")
+  return self.mergeOption(string, ls.getHover(filename, location), "getHover")
 
 method getInlayHints*(self: LanguageServerList, filename: string, selection: Selection): Future[Response[seq[language_server_base.InlayHint]]] {.async.} =
-  return mergeResponse(language_server_base.InlayHint, ls.getInlayHints(filename, selection), "getInlayHints")
+  return self.mergeResponse(language_server_base.InlayHint, ls.getInlayHints(filename, selection), "getInlayHints")
 
 method getDiagnostics*(self: LanguageServerList, filename: string): Future[Response[seq[lsp_types.Diagnostic]]] {.async.} =
-  return mergeResponse(lsp_types.Diagnostic, ls.getDiagnostics(filename), "getDiagnostics")
+  return self.mergeResponse(lsp_types.Diagnostic, ls.getDiagnostics(filename), "getDiagnostics")
 
 method getCompletionTriggerChars*(self: LanguageServerList): set[char] =
   for ls in self.languageServers:
     result.incl ls.getCompletionTriggerChars()
 
 method getCodeActions*(self: LanguageServerList, filename: string, selection: Selection, diagnostics: seq[lsp_types.Diagnostic]): Future[Response[lsp_types.CodeActionResponse]] {.async.} =
-  return mergeResponse(lsp_types.CodeActionResponseVariant, ls.getCodeActions(filename, selection, diagnostics), "getCodeActions")
+  return self.mergeResponse(lsp_types.CodeActionResponseVariant, ls.getCodeActions(filename, selection, diagnostics), "getCodeActions")
 
 method rename*(self: LanguageServerList, filename: string, position: Cursor, newName: string): Future[Response[seq[lsp_types.WorkspaceEdit]]] {.async.} =
-  return mergeResponse(lsp_types.WorkspaceEdit, ls.rename(filename, position, newName), "rename")
+  return self.mergeResponse(lsp_types.WorkspaceEdit, ls.rename(filename, position, newName), "rename")
 
 method executeCommand*(self: LanguageServerList, command: string, arguments: seq[JsonNode]): Future[Response[JsonNode]] {.async.} =
+  let timeout = self.mergeConfig.timeout.get()
   var futs = newSeq[Future[Response[JsonNode]]]()
   var futsTimeout = newSeq[Future[bool]]()
   for ls in self.languageServers:
     if ls.capabilities.executeCommandProvider.isSome and command in ls.capabilities.executeCommandProvider.get.commands:
       let fut = ls.executeCommand(command, arguments)
       futs.add fut
-      futsTimeout.add fut.withTimeout(1000.milliseconds)
+      futsTimeout.add fut.withTimeout(timeout.milliseconds)
 
   var res = errorResponse[JsonNode](0, "Command not found: " & command)